Address Parquet writer review comments
