+++ java/src/main/java/ai/rapids/cudf/Table.java
@@ -327,115 +327,6 @@ public final class Table implements AutoCloseable {
                                                       long dataSourceHandle) throws CudfException;
 
   /**
-   * Setup everything to write parquet formatted data to a file.
-   * @param columnNames       names that correspond to the table columns
-   * @param numChildren       Children of the top level
-   * @param flatNumChildren   flattened list of children per column
-   * @param nullable          true if the column can have nulls else false
-   * @param metadataKeys      Metadata key names to place in the Parquet file
-   * @param metadataValues    Metadata values corresponding to metadataKeys
-   * @param compression       native compression codec ID
-   * @param rowGroupSizeRows  max #rows in a row group
-   * @param rowGroupSizeBytes max #bytes in a row group
-   * @param maxDictionarySize maximum dictionary size in bytes (cuDF default 1 MiB)
-   * @param dictionaryPolicy  native dictionary policy ID
-   * @param statsFreq         native statistics frequency ID
-   * @param isInt96           true if timestamp type is int96
-   * @param precisions        precision list containing all the precisions of the decimal types in
-   *                          the columns
-   * @param isMapValues       true if a column is a map
-   * @param isBinaryValues    true if a column is a binary
-   * @param filename          local output path
-   * @return a handle that is used in later calls to writeParquetChunk and writeParquetEnd.
-   */
-  private static native long writeParquetFileBegin(String[] columnNames,
-                                                   int numChildren,
-                                                   int[] flatNumChildren,
-                                                   boolean[] nullable,
-                                                   String[] metadataKeys,
-                                                   String[] metadataValues,
-                                                   int compression,
-                                                   int rowGroupSizeRows,
-                                                   long rowGroupSizeBytes,
-                                                   long maxDictionarySize,
-                                                   int dictionaryPolicy,
-                                                   int statsFreq,
-                                                   boolean[] isInt96,
-                                                   int[] precisions,
-                                                   boolean[] isMapValues,
-                                                   boolean[] isBinaryValues,
-                                                   boolean[] hasParquetFieldIds,
-                                                   int[] parquetFieldIds,
-                                                   String filename) throws CudfException;
-
-  /**
-   * Setup everything to write parquet formatted data to a buffer.
-   * @param columnNames       names that correspond to the table columns
-   * @param numChildren       Children of the top level
-   * @param flatNumChildren   flattened list of children per column
-   * @param nullable          true if the column can have nulls else false
-   * @param metadataKeys      Metadata key names to place in the Parquet file
-   * @param metadataValues    Metadata values corresponding to metadataKeys
-   * @param compression       native compression codec ID
-   * @param rowGroupSizeRows  max #rows in a row group
-   * @param rowGroupSizeBytes max #bytes in a row group
-   * @param maxDictionarySize maximum dictionary size in bytes (cuDF default 1 MiB)
-   * @param dictionaryPolicy  native dictionary policy ID
-   * @param statsFreq         native statistics frequency ID
-   * @param isInt96           true if timestamp type is int96
-   * @param precisions        precision list containing all the precisions of the decimal types in
-   *                          the columns
-   * @param isMapValues       true if a column is a map
-   * @param isBinaryValues    true if a column is a binary
-   * @param consumer          consumer of host buffers produced.
-   * @return a handle that is used in later calls to writeParquetChunk and writeParquetEnd.
-   */
-  private static native long writeParquetBufferBegin(String[] columnNames,
-                                                     int numChildren,
-                                                     int[] flatNumChildren,
-                                                     boolean[] nullable,
-                                                     String[] metadataKeys,
-                                                     String[] metadataValues,
-                                                     int compression,
-                                                     int rowGroupSizeRows,
-                                                     long rowGroupSizeBytes,
-                                                     long maxDictionarySize,
-                                                     int dictionaryPolicy,
-                                                     int statsFreq,
-                                                     boolean[] isInt96,
-                                                     int[] precisions,
-                                                     boolean[] isMapValues,
-                                                     boolean[] isBinaryValues,
-                                                     boolean[] hasParquetFieldIds,
-                                                     int[] parquetFieldIds,
-                                                     HostBufferConsumer consumer,
-                                                     HostMemoryAllocator hostMemoryAllocator
-                                                     ) throws CudfException;
-
-  /**
-   * Write out a table to an open handle.
-   * @param handle the handle to the writer.
-   * @param table the table to write out.
-   * @param tableMemSize the size of the table in bytes to help with memory allocation.
-   */
-  private static native void writeParquetChunk(long handle, long table, long tableMemSize);
-
-  /**
-   * Finish writing out parquet.
-   * @param handle the handle.  Do not use again once this returns.
-   */
-  private static native void writeParquetEnd(long handle);
-
-  /**
-   * Finish writing out parquet and return its footer metadata.
-   * @param handle the handle. Do not use again once this returns.
-   * @param hostMemoryAllocator allocator for the returned host buffer
-   * @return a host buffer containing a metadata-only Parquet file
-   */
-  private static native HostMemoryBuffer writeParquetEndAndGetFooter(long handle,
-      HostMemoryAllocator hostMemoryAllocator);
-
-  /**
    * Read in ORC formatted data.
    * @param filterColumnNames name of the columns to read, or an empty array if we want to read
    *                          all of them
@@ -1687,127 +1578,6 @@ public final class Table implements AutoCloseable {
     }
   }
 
-  /** A chunked Parquet writer that can return its footer metadata when closed. */
-  public static final class ParquetTableWriter extends TableWriter {
-    private HostBufferConsumer consumer;
-    private final HostMemoryAllocator hostMemoryAllocator;
-
-    private ParquetTableWriter(ParquetWriterOptions options, File outputFile) {
-      super(writeParquetFileBegin(options.getFlatColumnNames(),
-          options.getTopLevelChildren(),
-          options.getFlatNumChildren(),
-          options.getFlatIsNullable(),
-          options.getMetadataKeys(),
-          options.getMetadataValues(),
-          options.getCompressionType().nativeId,
-          options.getRowGroupSizeRows(),
-          options.getRowGroupSizeBytes(),
-          options.getMaxDictionarySize(),
-          options.getDictionaryPolicy().nativeId,
-          options.getStatisticsFrequency().nativeId,
-          options.getFlatIsTimeTypeInt96(),
-          options.getFlatPrecision(),
-          options.getFlatIsMap(),
-          options.getFlatIsBinary(),
-          options.getFlatHasParquetFieldId(),
-          options.getFlatParquetFieldId(),
-          outputFile.getAbsolutePath()));
-      this.consumer = null;
-      this.hostMemoryAllocator = DefaultHostMemoryAllocator.get();
-    }
-
-    private ParquetTableWriter(ParquetWriterOptions options, HostBufferConsumer consumer,
-        HostMemoryAllocator hostMemoryAllocator) {
-      super(writeParquetBufferBegin(options.getFlatColumnNames(),
-          options.getTopLevelChildren(),
-          options.getFlatNumChildren(),
-          options.getFlatIsNullable(),
-          options.getMetadataKeys(),
-          options.getMetadataValues(),
-          options.getCompressionType().nativeId,
-          options.getRowGroupSizeRows(),
-          options.getRowGroupSizeBytes(),
-          options.getMaxDictionarySize(),
-          options.getDictionaryPolicy().nativeId,
-          options.getStatisticsFrequency().nativeId,
-          options.getFlatIsTimeTypeInt96(),
-          options.getFlatPrecision(),
-          options.getFlatIsMap(),
-          options.getFlatIsBinary(),
-          options.getFlatHasParquetFieldId(),
-          options.getFlatParquetFieldId(),
-          consumer, hostMemoryAllocator));
-      this.consumer = consumer;
-      this.hostMemoryAllocator = hostMemoryAllocator;
-    }
-
-    @Override
-    public void write(Table table) {
-      if (writerHandle == 0) {
-        throw new IllegalStateException("Writer was already closed");
-      }
-      writeParquetChunk(writerHandle, table.nativeHandle, table.getDeviceMemorySize());
-    }
-
-    @Override
-    public void close() throws CudfException {
-      if (writerHandle != 0) {
-        finish(false);
-      }
-    }
-
-    /**
-     * Finish writing and return the Parquet footer metadata.
-     *
-     * <p>The returned buffer is a metadata-only Parquet file containing the leading {@code PAR1}
-     * magic, the serialized file metadata, the footer length, and the trailing {@code PAR1}
-     * magic. The caller must close the returned buffer.
-     *
-     * @return an owned host buffer containing the Parquet footer metadata
-     * @throws IllegalStateException if the writer was already closed
-     */
-    public HostMemoryBuffer closeAndGetFooter() throws CudfException {
-      if (writerHandle == 0) {
-        throw new IllegalStateException("Writer was already closed");
-      }
-      return finish(true);
-    }
-
-    private HostMemoryBuffer finish(boolean returnFooter) {
-      long handle = writerHandle;
-      writerHandle = 0;
-      HostBufferConsumer currentConsumer = consumer;
-      consumer = null;
-      HostMemoryBuffer footer = null;
-      try {
-        if (returnFooter) {
-          footer = writeParquetEndAndGetFooter(handle, hostMemoryAllocator);
-        } else {
-          writeParquetEnd(handle);
-        }
-      } catch (RuntimeException | Error e) {
-        if (currentConsumer != null) {
-          try {
-            currentConsumer.done();
-          } catch (RuntimeException | Error doneError) {
-            e.addSuppressed(doneError);
-          }
-        }
-        throw e;
-      }
-
-      try {
-        if (currentConsumer != null) {
-          currentConsumer.done();
-        }
-      } catch (RuntimeException | Error e) {
-        CleanupHelpers.closeAndSuppress(footer, e);
-        throw e;
-      }
-      return footer;
-    }
-  }
-
   /**
    * Get a table writer to write parquet data to a file.
    * @param options the parquet writer options.
@@ -1875,7 +1645,7 @@ public final class Table implements AutoCloseable {
         for (ColumnView cv : columnViews) {
           total += cv.getDeviceMemorySize();
         }
-        writeParquetChunk(writer.writerHandle, nativeHandle, total);
+        writer.write(nativeHandle, total);
       }
     } finally {
       deleteCudfTable(nativeHandle);
